+++ src/main/java/com/watermelon/dateapp/api/dto/CreateUserRequest.java
@@ -1,4 +1,4 @@
-package com.watermelon.dateapp.controller;
+package com.watermelon.dateapp.api.dto;
 
 import jakarta.validation.constraints.NotBlank;
